correção na interface auxiliar

+++ src/interfaces/GerenciadorFilmes_IF.java
@@ -4,9 +4,9 @@ public interface GerenciadorFilmes_IF {
 
     void inserir(Filme_IF elemento);
 
-    Filme_IF buscar(int id) throws Exception;
+    Filme_IF buscar(long id) throws Exception;
 
-    Filme_IF remover(int id) throws Exception;
+    Filme_IF remover(long id) throws Exception;
 
     Filme_IF[] ordenar() throws Exception;
 }
